requirements.txt
Pillow==6.1.0
argon2-cffi==19.2.0
gunicorn==19.7.1
django-heroku==0.3.1
